+++ src/navigation/BottomTabNavigator.js
@@ -26,12 +26,10 @@ import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
 
 import HomeScreen from '../screens/home/HomeScreen';
 import ListMainScreen from '../screens/list/ListMainScreen';
-import ListDetailScreen from '../screens/list/ListDetailScreen';
 import MapScreen from '../screens/map/MapScreen';
 import MypageScreen from '../screens/mypage/MypageScreen';
 import SearchScreen from '../screens/map/SearchScreen';
-import StoreDetailScreen from '../screens/detail/StoreDetailScreen';
-import ReviewWriteScreen from '../screens/detail/ReviewWriteScreen';
+import UserDataChangeScreen from '../screens/mypage/UserDataChangeScreen';
 
 const BottomTab = createBottomTabNavigator();
 const iconSize = 35;
@@ -113,7 +111,6 @@ function ListNavigator() {
         cardStyleInterpolator: customCardStyleInterpolator,
       }}>
       <ListStack.Screen name="ListMain" component={ListMainScreen} />
-      <ListStack.Screen name="ListDetail" component={ListDetailScreen} />
     </ListStack.Navigator>
   );
 }
@@ -131,6 +128,10 @@ function MypageNavigator() {
         cardStyleInterpolator: customCardStyleInterpolator,
       }}>
       <MypageStack.Screen name="Mypage" component={MypageScreen} />
+      <MypageStack.Screen
+        name="UserDataChange"
+        component={UserDataChangeScreen}
+      />
     </MypageStack.Navigator>
   );
 }
