+++ src/screens/mypage/UserDataChangeScreen.js
@@ -1,0 +1,31 @@
+import React, {useState, useCallback, useEffect} from 'react';
+import {View, Text, SafeAreaView, ScrollView, StyleSheet} from 'react-native';
+import {
+  COLOR_WHITE,
+  COLOR_BACKGROUND,
+  COLOR_GRAY,
+  COLOR_PRIMARY,
+} from '../../assets/color';
+import AnimatedButton from '../../components/AnimationButton';
+import {useNavigation} from '@react-navigation/native';
+import Header from '../../components/Header';
+
+export default function UserDataChangeScreen() {
+  const navigation = useNavigation();
+
+  return (
+    <>
+      <Header title={'내 정보 수정'} isBackButton={true} />
+      <View style={styles.entire}></View>
+    </>
+  );
+}
+
+const styles = StyleSheet.create({
+  entire: {
+    flex: 1,
+    backgroundColor: COLOR_BACKGROUND,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+});
